+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - UNRELEASED
+## [1.0.1] - 2026-02-02
 
-Initial release.
+### Added
+
+- Create default example config, if none exists already.
+- Create default `swapper.json` to easily migrate from Block Swap.
+
+## [1.0.0] - 2026-01-31
+
+* Initial release.